+++ build_runner/lib/src/daemon/daemon_builder.dart
@@ -258,7 +258,6 @@ class BuildRunnerDaemonBuilder implements DaemonBuilder {
     var buildOptions = await BuildOptions.create(
       packageGraph: packageGraph,
       reader: daemonEnvironment.reader,
-      deleteFilesByDefault: daemonOptions.deleteFilesByDefault,
       overrideBuildConfig: overrideBuildConfig,
       skipBuildScriptCheck: daemonOptions.skipBuildScriptCheck,
       enableLowResourcesMode: daemonOptions.enableLowResourcesMode,